requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.21
d4rl==1.1
decorator==4.4.2
dm-control==0.0.349668059
dm-env==1.4
dm-tree==0.1.5
dmc2gym==1.0.0
fasteners==0.16
future==0.18.2
gast==0.2.2
glfw==2.1.0
google-pasta==0.2.0
grpcio==1.35.0
gym==0.18.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labmaze==1.0.3
llvmlite==0.36.0
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.4
mjrl==1.0.0
moviepy==1.0.3
mpi4py==3.0.3
mujoco-py==2.0.2.13
networkx==2.5
numba==0.53.1
numpy==1.20.1
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.2.4
Pillow==7.2.0
proglog==0.1.9
protobuf==3.14.0
pybullet==3.1.2
pycparser==2.20
pygame==2.0.1
pyglet==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
robosuite==1.2.1
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
tensorboard==1.15.0
tensorboardX==2.1
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2021.2.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
